Add files via upload
This program deals with the complexities of SQLite database that can occur or usually occurs in any app designing which includes the usage of database for multiple aspects. Also this program is a demonstration to store ones personal information in an organised manner. also it can store multiple subjects details, so you can add you family members information too.