Add files via upload
It is a music player application which works on voice command of the user. It basically converts you speech into text and stores into an arraylist and then checks whether that song is present in the list or not and then plays it. You can control the application (forward, backward, track the timing) as well. It has its own raw folder of songs so that everytime it plays a new song and let the user to discover a new taste and new genre of music